Wireframe created
The WireframeMain is the detailed version of the game wireframe. The mockup design is more of a look at how the car will fly through the air and the style of game that it is. Note that the game is in 3d but the path is fixed that the car must travel down only the speed and boost variables are controllable, no left and right hand turning.